requirements.txt
alembic==0.9.6
asn1crypto==0.24.0
attrs==17.3.0
aws-xray-sdk==0.95
bcrypt==3.1.4
boto==2.48.0
boto3==1.5.4
botocore==1.8.18
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
colorama==0.3.9
cookies==2.2.1
cryptography==2.1.4
docker==2.7.0
docker-pycreds==0.2.1
docutils==0.14
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
flask-marshmallow==0.8.0
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.0
idna==2.6
isodate==0.6.0
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
jsondiff==1.1.1
jsonpickle==0.9.5
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.15.0
mock==2.0.0
moto==1.1.25
pbr==3.1.1
pluggy==0.6.0
py==1.5.2
pyaml==17.10.0
pycparser==2.18
pytest==3.3.1
pytest-flask==0.10.0
python-dateutil==2.6.1
python-editor==1.0.3
python-slugify==1.2.4
pytz==2017.3
PyYAML==3.12
requests==2.18.4
s3transfer==0.1.12
six==1.11.0
SQLAlchemy==1.1.14
SQLAlchemy-Utils==0.32.17
Unidecode==0.4.21
urllib3==1.22
websocket-client==0.44.0
Werkzeug==0.12.2
wrapt==1.10.11
xmltodict==0.11.0
